+++ cli/coverage-node.js
@@ -1,9 +1,14 @@
 #!/usr/bin/env node
 
 'use strict'
+
+const { spawn } = require('child_process')
+const kleur = require('kleur')
 const analyseCoverage = require('../lib/analyseCoverage')
+const childProcessPromise = require('../lib/childProcessPromise')
+const coverageSupported = require('../lib/coverageSupported')
+const minNodeVersion = require('../lib/coverageSupportedMinNodeVersion')
 const errorConsole = require('../lib/errorConsole')
-const nodeWithCoverage = require('../lib/nodeWithCoverage')
 const reportCoverage = require('../lib/reportCoverage')
 const tempDirOperation = require('../lib/tempDirOperation')
 
@@ -22,23 +27,43 @@ async function coverageNode() {
   try {
     if (!nodeArgs.length) throw new Error('Node.js CLI arguments are required.')
 
-    await tempDirOperation(async tempDirPath => {
-      const exitCode = await nodeWithCoverage(tempDirPath, nodeArgs)
+    // eslint-disable-next-line curly
+    if (coverageSupported) {
+      await tempDirOperation(async tempDirPath => {
+        const { exitCode } = await childProcessPromise(
+          spawn('node', nodeArgs, {
+            stdio: 'inherit',
+            env: { ...process.env, NODE_V8_COVERAGE: tempDirPath }
+          })
+        )
 
-      // Only show a code coverage report if the Node.js script didn’t error,
-      // to reduce distraction from the priority to solve errors.
-      if (exitCode === 0) {
-        const analysis = await analyseCoverage(tempDirPath)
-        reportCoverage(analysis)
-        if (analysis.uncovered.length) process.exitCode = 1
-      }
+        // Only show a code coverage report if the Node.js script didn’t error,
+        // to reduce distraction from the priority to solve errors.
+        if (exitCode === 0) {
+          const analysis = await analyseCoverage(tempDirPath)
+          reportCoverage(analysis)
+          if (analysis.uncovered.length) process.exitCode = 1
+        } else process.exitCode = exitCode
+      })
+      // coverage ignore next line
+    } else {
+      const { exitCode } = await childProcessPromise(
+        spawn('node', nodeArgs, { stdio: 'inherit' })
+      )
 
-      // The error exit code of the spawned subprocess must be manually applied
-      // to this parent process.
+      if (exitCode === 0)
+        console.info(
+          `\n${kleur.yellow(
+            `Skipped code coverage as Node.js is ${process.version}, v${minNodeVersion.major}.${minNodeVersion.minor}.${minNodeVersion.patch}+ is supported.`
+          )}\n`
+        )
       else process.exitCode = exitCode
-    })
+    }
   } catch (error) {
-    errorConsole.group('Error running Node.js with coverage:')
+    errorConsole.group(
+      // coverage ignore next line
+      `Error running Node.js${coverageSupported ? '  with coverage' : ''}:`
+    )
     errorConsole.error(error)
     errorConsole.groupEnd()
     process.exitCode = 1